refactor(kube-api-rewriter): more rewrites for AdmissionReview
- Rewrite metadata in patches for mutating responses.
- Full restore for object and oldObject in AdmisstionReview request.
- Move admission request test to admission_review_test.go.
- Add webhookRules to createTestRewriter.
- Allow rewrite for JSON responses for non application/json Content-Type. E.g. text/plain responses from webhooks.
